hatebert/requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
blis==0.4.1
boto3==1.11.16
botocore==1.14.16
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.0
cymem==2.0.3
dataclasses==0.7
decorator==4.4.1
docutils==0.15.2
emoji==0.5.4
filelock==3.0.12
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
jmespath==0.9.4
joblib==0.14.1
keras==2.3.1
keras-applications==1.0.8
keras-preprocessing==1.1.0
mako==1.1.1
markdown==3.2.1
markupsafe==1.1.1
murmurhash==1.0.2
ninja==1.9.0
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
pillow==7.0.0
pip==20.0.2
plac==1.1.3
preshed==3.0.2
prettytable==0.7.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytools==2020.1
pyyaml==5.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
sentence-transformers==0.2.5
sentencepiece==0.1.85
setuptools==45.2.0
six==1.14.0
sklearn==0.0
spacy==2.2.3
srsly==1.0.1
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
thinc==7.3.1
tokenizers==0.5.2
torchvision==0.5.0
tqdm==4.42.1
transformers==2.3.0
urllib3==1.25.8
wasabi==0.6.0
werkzeug==1.0.0
wheel==0.34.2
wrapt==1.11.2
zipp==2.2.0